Convert RoomLinkValidator to TypeScript
